scripts/requirements.txt
# Python dependencies for PAM Multi-Vendor Lab
# Version 2.0 - January 2026
# Author: Dominic M. Hoang

# ============================================
# HashiCorp Vault
# ============================================
hvac>=2.1.0

# ============================================
# Delinea Secret Server (NEW in v2.0)
# ============================================
python-tss-sdk>=1.2.0

# ============================================
# AWS Secrets Manager
# ============================================
boto3>=1.34.0
botocore>=1.34.0

# ============================================
# HTTP & API
# ============================================
requests>=2.31.0

# ============================================
# Configuration
# ============================================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# CLI & Terminal UI
# ============================================
click>=8.1.0
rich>=13.7.0
colorama>=0.4.6
tabulate>=0.9.0

# ============================================
# Database Connectivity
# ============================================
psycopg2-binary>=2.9.0
PyMySQL>=1.1.0

# ============================================
# SSH Connectivity
# ============================================
paramiko>=3.4.0

# ============================================
# JSON & YAML
# ============================================
python-json-logger>=2.0.0
pyyaml>=6.0.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
responses>=0.24.0

# ============================================
# Code Quality
# ============================================
black>=23.12.0
flake8>=7.0.0
pylint>=3.0.0
mypy>=1.8.0

# ============================================
# Optional: Slack Alerts
# ============================================
# slack-sdk>=3.23.0

# ============================================
# Optional: PDF Reports
# ============================================
# reportlab>=4.0.0
